requirements.txt
streamlit
Pillow
plotly
pandas
numpy